style: Move length delimiter encoding to separate module
Leave a alias in `lib.rs` to prevent a breaking change.
